% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{.correlation_to_covariance}
\alias{.correlation_to_covariance}
\title{Convert correlation to covariance matrix}
\usage{
.correlation_to_covariance(corr_matrix, std_vector)
}
\description{
Convert correlation to covariance matrix
}
\keyword{internal}
