% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{.extract_features}
\alias{.extract_features}
\title{Extract histogram features from standardized data}
\usage{
.extract_features(
  data,
  ndim,
  feature_grid_range = c(-4, 4),
  feature_grid_nbins = 100
)
}
\description{
Extract histogram features from standardized data
}
\keyword{internal}
